backend/requirements.txt
fastapi==0.103.2
uvicorn[standard]==0.32.1
sqlmodel==0.0.8
SQLAlchemy==1.4.41
pydantic==1.10.13

psycopg2-binary==2.9.10
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
cloudinary==1.41.0
httpx==0.25.2